font,text: add font.(*Face).LineSize and shrink text.String
This change first adds font.(*Face).LineSize, essentially making some of
the calculations that the text package does part of the public API.
Then, it uses this new API to simplify the text package implementation.

While we're here simplify text.String to not track the direction
explicitly. We can track it implicitly by looking at the direction of
the first segment, since all segments must have the same direction.